Add rendition helper and move SCORE logic into getStartCodecTier
