+++ src/controller/abr-controller.ts
@@ -2,8 +2,12 @@ import EwmaBandWidthEstimator from '../utils/ewma-bandwidth-estimator';
 import { Events } from '../events';
 import { ErrorDetails } from '../errors';
 import { PlaylistLevelType } from '../types/loader';
-import { codecsSetSelectionPreferenceValue } from '../utils/codecs';
 import { logger } from '../utils/logger';
+import {
+  getCodecTiers,
+  getStartCodecTier,
+  type CodecSetTier,
+} from './rendition-helper';
 import type { Fragment } from '../loader/fragment';
 import type { Part } from '../loader/fragment';
 import type { Level, VideoRange } from '../types/level';
@@ -19,28 +23,6 @@ import type {
   ErrorData,
 } from '../types/events';
 import type { AbrComponentAPI } from '../types/component-api';
-
-type CodecSetTier = {
-  minBitrate: number;
-  minHeight: number;
-  minFramerate: number;
-  videoRanges: Record<string, number>;
-  channels: Record<string, number>;
-  hasDefaultAudio: boolean;
-  fragmentError: number;
-};
-
-type AudioTrackGroup = {
-  channels: Record<string, number>;
-  hasDefault: boolean;
-  hasAutoSelect: boolean;
-};
-type StartParameters = {
-  codecSet: string | undefined;
-  videoRange: VideoRange | undefined;
-  minFramerate: number;
-  minBitrate: number;
-};
 
 class AbrController implements AbrComponentAPI {
   protected hls: Hls;
@@ -620,7 +602,7 @@ class AbrController implements AbrComponentAPI {
     const selectionBaseLevel =
       lastLoadedFragLevel === -1 ? this.hls.firstLevel : lastLoadedFragLevel;
     const { fragCurrent, partCurrent } = this;
-    const { levels, loadLevel } = this.hls;
+    const { levels, allAudioTracks, loadLevel } = this.hls;
     const level: Level | undefined = levels[selectionBaseLevel];
     const live = !!level?.details?.live;
     const firstSelection = loadLevel === -1 || lastLoadedFragLevel === -1;
@@ -628,13 +610,16 @@ class AbrController implements AbrComponentAPI {
     let currentVideoRange: VideoRange | undefined = 'SDR';
     let currentFrameRate = level?.frameRate || 0;
     if (firstSelection) {
-      const { codecSet, videoRange, minFramerate, minBitrate } =
-        this.getStartCodecTier(
-          currentVideoRange,
-          currentBw,
+      const codecTiers =
+        this.codecTiers ||
+        (this.codecTiers = getCodecTiers(
+          levels,
+          allAudioTracks,
           minAutoLevel,
           maxAutoLevel
-        );
+        ));
+      const { codecSet, videoRange, minFramerate, minBitrate } =
+        getStartCodecTier(codecTiers, currentVideoRange, currentBw);
       currentCodecSet = codecSet;
       currentVideoRange = videoRange;
       currentFrameRate = minFramerate;
@@ -756,192 +741,6 @@ class AbrController implements AbrComponentAPI {
     return -1;
   }
 
-  private getStartCodecTier(
-    videoRange: VideoRange | undefined,
-    currentBw: number,
-    minAutoLevel: number,
-    maxAutoLevel: number
-  ): StartParameters {
-    const codecTiers = this.getCodecTiers(minAutoLevel, maxAutoLevel);
-    const codecSets = Object.keys(codecTiers);
-    // Use first level set to determine stereo, and minimum resolution and framerate
-    let hasStereo = true;
-    let hasCurrentVideoRange = false;
-    let minHeight = Infinity;
-    let minFramerate = Infinity;
-    let minBitrate = Infinity;
-    for (let i = codecSets.length; i--; ) {
-      const tier = codecTiers[codecSets[i]];
-      hasStereo = tier.channels[2] > 0;
-      minHeight = Math.min(minHeight, tier.minHeight);
-      minFramerate = Math.min(minFramerate, tier.minFramerate);
-      minBitrate = Math.min(minBitrate, tier.minBitrate);
-      if (videoRange) {
-        hasCurrentVideoRange ||= tier.videoRanges[videoRange] > 0;
-      }
-    }
-    minHeight = Number.isFinite(minHeight) ? minHeight : 0;
-    minFramerate = Number.isFinite(minFramerate) ? minFramerate : 0;
-    const maxHeight = Math.max(1080, minHeight);
-    const maxFramerate = Math.max(30, minFramerate);
-    minBitrate = Number.isFinite(minBitrate) ? minBitrate : currentBw;
-    currentBw = Math.max(minBitrate, currentBw);
-    // If there are no SDR variants, set currentVideoRange to undefined
-    if (!hasCurrentVideoRange) {
-      videoRange = undefined;
-    }
-    const codecSet = codecSets.reduce(
-      (selected: string | undefined, candidate: string) => {
-        // Remove candiates which do not meet bitrate, default audio, stereo, 1080p or lower, 30fps or lower, or SDR if present
-        const candidateTier = codecTiers[candidate];
-        if (candidate === selected) {
-          return selected;
-        }
-        if (candidateTier.minBitrate > currentBw) {
-          logStartCodecCandidateIgnored(
-            candidate,
-            `min bitrate of ${candidateTier.minBitrate} > current estimate of ${currentBw}`
-          );
-          return selected;
-        }
-        if (!candidateTier.hasDefaultAudio) {
-          logStartCodecCandidateIgnored(
-            candidate,
-            `no renditions with default or auto-select sound found`
-          );
-          return selected;
-        }
-        if (hasStereo && candidateTier.channels['2'] === 0) {
-          logStartCodecCandidateIgnored(
-            candidate,
-            `no renditions with stereo sound found`
-          );
-          return selected;
-        }
-        if (candidateTier.minHeight > maxHeight) {
-          logStartCodecCandidateIgnored(
-            candidate,
-            `min resolution of ${candidateTier.minHeight} > maximum of ${maxHeight}`
-          );
-          return selected;
-        }
-        if (candidateTier.minFramerate > maxFramerate) {
-          logStartCodecCandidateIgnored(
-            candidate,
-            `min framerate of ${candidateTier.minFramerate} > maximum of ${maxFramerate}`
-          );
-          return selected;
-        }
-        if (videoRange && candidateTier.videoRanges[videoRange] === 0) {
-          logStartCodecCandidateIgnored(
-            candidate,
-            `no variants with VIDEO-RANGE of ${videoRange} found`
-          );
-          return selected;
-        }
-        // Remove candiates with less preferred codecs or more errors
-        if (
-          selected &&
-          (codecsSetSelectionPreferenceValue(candidate) >=
-            codecsSetSelectionPreferenceValue(selected) ||
-            candidateTier.fragmentError > codecTiers[selected].fragmentError)
-        ) {
-          return selected;
-        }
-        return candidate;
-      },
-      undefined
-    );
-    return {
-      codecSet,
-      videoRange,
-      minFramerate,
-      minBitrate,
-    };
-  }
-
-  private getCodecTiers(
-    minAutoLevel: number,
-    maxAutoLevel: number
-  ): Record<string, CodecSetTier> {
-    let codecTiers = this.codecTiers;
-    if (codecTiers === null) {
-      let hasDefaultAudio = false;
-      let hasAutoSelectAudio = false;
-      const audioTracksByGroup = this.hls.allAudioTracks.reduce(
-        (trackGroups: Record<string, AudioTrackGroup>, track) => {
-          let trackGroup = trackGroups[track.groupId];
-          if (!trackGroup) {
-            trackGroup = trackGroups[track.groupId] = {
-              channels: { 2: 0 },
-              hasDefault: false,
-              hasAutoSelect: false,
-            };
-          }
-          const channels = track.attrs.CHANNELS;
-          trackGroup.channels[channels || 2] =
-            (trackGroup.channels[channels || 2] || 0) + 1;
-          trackGroup.hasDefault = trackGroup.hasDefault || track.default;
-          trackGroup.hasAutoSelect =
-            trackGroup.hasAutoSelect || track.autoselect;
-          if (trackGroup.hasDefault) {
-            hasDefaultAudio = true;
-          }
-          if (trackGroup.hasAutoSelect) {
-            hasAutoSelectAudio = true;
-          }
-          return trackGroups;
-        },
-        {}
-      );
-
-      codecTiers = this.codecTiers = this.hls.levels
-        .slice(minAutoLevel, maxAutoLevel + 1)
-        .reduce((tiers: Record<string, CodecSetTier>, level) => {
-          if (!level.codecSet) {
-            return tiers;
-          }
-          const audioGroup = level.audioGroupId
-            ? audioTracksByGroup[level.audioGroupId]
-            : null;
-          let tier = tiers[level.codecSet];
-          if (!tier) {
-            tiers[level.codecSet] = tier = {
-              minBitrate: Infinity,
-              minHeight: Infinity,
-              minFramerate: Infinity,
-              videoRanges: { SDR: 0 },
-              channels: { 2: 0 },
-              hasDefaultAudio: !audioGroup,
-              fragmentError: 0,
-            };
-          }
-          tier.minBitrate = Math.min(tier.minBitrate, level.bitrate);
-          const lesserWidthOrHeight = Math.min(level.height, level.width);
-          tier.minHeight = Math.min(tier.minHeight, lesserWidthOrHeight);
-          tier.minFramerate = Math.min(tier.minFramerate, level.frameRate);
-          tier.fragmentError += level.fragmentError;
-          tier.videoRanges[level.videoRange] =
-            (tier.videoRanges[level.videoRange] || 0) + 1;
-          if (audioGroup) {
-            // Default audio is any group with DEFAULT=YES, or if missing then any group with AUTOSELECT=YES, or all variants
-            tier.hasDefaultAudio =
-              tier.hasDefaultAudio || hasDefaultAudio
-                ? audioGroup.hasDefault
-                : audioGroup.hasAutoSelect ||
-                  (!hasDefaultAudio && !hasAutoSelectAudio);
-            Object.keys(audioGroup.channels).forEach((channels) => {
-              tier.channels[channels] =
-                (tier.channels[channels] || 0) + audioGroup.channels[channels];
-            });
-          }
-
-          return tiers;
-        }, {});
-    }
-    return codecTiers;
-  }
-
   set nextAutoLevel(nextLevel: number) {
     const value = Math.max(this.hls.minAutoLevel, nextLevel);
     if (this._nextAutoLevel != value) {
@@ -949,12 +748,6 @@ class AbrController implements AbrComponentAPI {
       this._nextAutoLevel = value;
     }
   }
-}
-
-function logStartCodecCandidateIgnored(codeSet: string, reason: string) {
-  logger.log(
-    `[abr] start candidates with "${codeSet}" ignored because ${reason}`
-  );
 }
 
 export default AbrController;
